cleanup: cleaned and segmented out files
