+++ src/pages/Account/Account.tsx
@@ -1,7 +1,5 @@
 import { FC, useState } from 'react'
-import styled from 'styled-components'
-import tw from 'twin.macro'
-import 'styled-components/macro'
+
 import AccountOverview from './components/AccountOverview'
 import Sidebar from './components/Sidebar'
 import DepositWithdrawHistory from './components/DepositWithdaw'
@@ -15,26 +13,20 @@ import MobileDepositWithdrawHistory from './components/Mobile/DepositWithdaw'
 import MobileTrades from './components/Mobile/Trades'
 import MobileFundingHistory from './components/Mobile/FundingHistory'
 
-const DESKTOPWRAPPER = styled.div`
-  ${tw`flex flex-row flex-nowrap`}
-`
-const MOBILEWRAPPER = styled.div`
-  ${tw`flex flex-col `}
-`
 const Account: FC = () => {
   const [selected, setSelected] = useState(0)
   const [selectedMenuItem, setSelectedMenuItem] = useState<'Deposits' | 'Trades' | 'Funding'>('Deposits')
   return !checkMobile() ? (
-    <DESKTOPWRAPPER className="no-scrollbar">
+    <div className="no-scrollbar flex flex-row flex-nowrap">
       <Sidebar selected={selected} setSelected={setSelected} />
       {selected === 0 && <AccountOverview />}
       {selected === 2 && <DepositWithdrawHistory />}
       {selected === 3 && <Trades />}
       {selected === 4 && <FundingHistory />}
       {selected === 5 && <Liquidations />}
-    </DESKTOPWRAPPER>
+    </div>
   ) : (
-    <MOBILEWRAPPER>
+    <div className={'flex flex-col'}>
       <MobileNav
         selected={selected}
         setSelected={setSelected}
@@ -45,7 +37,7 @@ const Account: FC = () => {
       {selected == 1 && selectedMenuItem === 'Deposits' && <MobileDepositWithdrawHistory />}
       {selected == 1 && selectedMenuItem === 'Trades' && <MobileTrades />}
       {selected == 1 && selectedMenuItem === 'Funding' && <MobileFundingHistory />}
-    </MOBILEWRAPPER>
+    </div>
   )
 }
 export default Account
